requirements.txt
asn1crypto==0.24.0
boto3==1.17.112
botocore==1.20.112
click==7.1.2
cryptography==2.1.4
Django==1.11.29
django-multiselectfield==0.1.12
django-storages==1.9.1
djangorestframework==3.9.4
enum34==1.1.6
example==0.1.0
Flask==1.1.4
futures==3.3.0
idna==2.6
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.1
mysql-connector-python==2.1.6
mysql-utilities==1.6.4
paramiko==2.0.0
pexpect==4.2.1
pyasn1==0.4.2
pycrypto==2.6.1
pygobject==3.26.1
PyMySQL==0.10.1
pyodbc==4.0.17
pysqlite==2.7.0
python-dateutil==2.8.2
pytz==2021.3
pyxdg==0.25
s3transfer==0.4.2
SecretStorage==2.3.1
six==1.11.0
urllib3==1.26.7
Werkzeug==1.0.1
